move node events component
